Move keyboard and mouse code to a common directory
