+++ src/hardware/input/mouse_common.cpp
