+++ src/hardware/input/mouse_config.h
