+++ src/hardware/input/mouse_manymouse.cpp
