+++ src/hardware/input/mouse_queue.h
