+++ src/hardware/input/mouseif_virtual_machines.cpp
