+++ src/hardware/meson.build
@@ -1,15 +1,16 @@
 libhardware_sources = files(
+    'input/keyboard.cpp',
+    'input/mouse.cpp',
+    'input/mouse_common.cpp',
+    'input/mouse_config.cpp',
+    'input/mouse_interfaces.cpp',
+    'input/mouse_manymouse.cpp',
+    'input/mouse_queue.cpp',
+    'input/mouseif_dos_driver.cpp',
+    'input/mouseif_ps2_bios.cpp',
+    'input/mouseif_virtual_machines.cpp',
     'mame/saa1099.cpp',
     'mame/sn76496.cpp',
-    'mouse/mouse.cpp',
-    'mouse/mouse_common.cpp',
-    'mouse/mouse_config.cpp',
-    'mouse/mouse_interfaces.cpp',
-    'mouse/mouse_manymouse.cpp',
-    'mouse/mouse_queue.cpp',
-    'mouse/mouseif_dos_driver.cpp',
-    'mouse/mouseif_ps2_bios.cpp',
-    'mouse/mouseif_virtual_machines.cpp',
     'serialport/directserial.cpp',
     'serialport/libserial.cpp',
     'serialport/misc_util.cpp',
@@ -35,7 +36,6 @@ libhardware_sources = files(
     'ipx.cpp',
     'ipxserver.cpp',
     'joystick.cpp',
-    'keyboard.cpp',
     'lpt_dac.cpp',
     'memory.cpp',
     'mixer.cpp',
