+++ src/hardware/serialport/serialmouse.cpp
@@ -1,5 +1,5 @@
 /*
- *  Copyright (C) 2022-2022  The DOSBox Staging Team
+ *  Copyright (C) 2022-2023  The DOSBox Staging Team
  *  Copyright (C) 2002-2021  The DOSBox Team
  *
  *  This program is free software; you can redistribute it and/or modify
@@ -30,7 +30,7 @@
 #include "checks.h"
 #include "math_utils.h"
 
-#include "../mouse/mouse_interfaces.h"
+#include "../input/mouse_interfaces.h"
 
 CHECK_NARROWING();
