+++ src/hardware/serialport/serialmouse.h
@@ -1,5 +1,5 @@
 /*
- *  Copyright (C) 2022-2022  The DOSBox Staging Team
+ *  Copyright (C) 2022-2023  The DOSBox Staging Team
  *  Copyright (C) 2002-2021  The DOSBox Team
  *
  *  This program is free software; you can redistribute it and/or modify
@@ -22,7 +22,7 @@
 
 #include "serialport.h"
 
-#include "../mouse/mouse_config.h"
+#include "../input/mouse_config.h"
 
 class CSerialMouse : public CSerial {
 public:
