+++ vs/dosbox.vcxproj
@@ -583,26 +583,26 @@ IF %ERRORLEVEL% LSS 8 SET ERRORLEVEL = 0</Command>
     <ClCompile Include="..\src\hardware\hardware.cpp" />
     <ClCompile Include="..\src\hardware\ide.cpp" />
     <ClCompile Include="..\src\hardware\innovation.cpp" />
+    <ClCompile Include="..\src\hardware\input\keyboard.cpp" />
+    <ClCompile Include="..\src\hardware\input\mouse.cpp" />
+    <ClCompile Include="..\src\hardware\input\mouse_common.cpp" />
+    <ClCompile Include="..\src\hardware\input\mouse_config.cpp" />
+    <ClCompile Include="..\src\hardware\input\mouse_interfaces.cpp" />
+    <ClCompile Include="..\src\hardware\input\mouse_manymouse.cpp" />
+    <ClCompile Include="..\src\hardware\input\mouse_queue.cpp" />
+    <ClCompile Include="..\src\hardware\input\mouseif_dos_driver.cpp" />
+    <ClCompile Include="..\src\hardware\input\mouseif_ps2_bios.cpp" />
+    <ClCompile Include="..\src\hardware\input\mouseif_virtual_machines.cpp" />
     <ClCompile Include="..\src\hardware\iohandler.cpp" />
     <ClCompile Include="..\src\hardware\iohandler_containers.cpp" />
     <ClCompile Include="..\src\hardware\ipx.cpp" />
     <ClCompile Include="..\src\hardware\ipxserver.cpp" />
     <ClCompile Include="..\src\hardware\joystick.cpp" />
-    <ClCompile Include="..\src\hardware\keyboard.cpp" />
     <ClCompile Include="..\src\hardware\lpt_dac.cpp" />
     <ClCompile Include="..\src\hardware\mame\saa1099.cpp" />
     <ClCompile Include="..\src\hardware\mame\sn76496.cpp" />
     <ClCompile Include="..\src\hardware\memory.cpp" />
     <ClCompile Include="..\src\hardware\mixer.cpp" />
-    <ClCompile Include="..\src\hardware\mouse\mouse.cpp" />
-    <ClCompile Include="..\src\hardware\mouse\mouse_common.cpp" />
-    <ClCompile Include="..\src\hardware\mouse\mouse_config.cpp" />
-    <ClCompile Include="..\src\hardware\mouse\mouse_interfaces.cpp" />
-    <ClCompile Include="..\src\hardware\mouse\mouse_manymouse.cpp" />
-    <ClCompile Include="..\src\hardware\mouse\mouse_queue.cpp" />
-    <ClCompile Include="..\src\hardware\mouse\mouseif_dos_driver.cpp" />
-    <ClCompile Include="..\src\hardware\mouse\mouseif_ps2_bios.cpp" />
-    <ClCompile Include="..\src\hardware\mouse\mouseif_virtual_machines.cpp" />
     <ClCompile Include="..\src\hardware\mpu401.cpp" />
     <ClCompile Include="..\src\hardware\opl.cpp" />
     <ClCompile Include="..\src\hardware\pci_bus.cpp" />
@@ -825,14 +825,14 @@ IF %ERRORLEVEL% LSS 8 SET ERRORLEVEL = 0</Command>
     <ClInclude Include="..\src\gui\render_templates.h" />
     <ClInclude Include="..\src\hardware\font-switch.h" />
     <ClInclude Include="..\src\hardware\gameblaster.h" />
+    <ClInclude Include="..\src\hardware\input\mouse_common.h" />
+    <ClInclude Include="..\src\hardware\input\mouse_config.h" />
+    <ClInclude Include="..\src\hardware\input\mouse_interfaces.h" />
+    <ClInclude Include="..\src\hardware\input\mouse_manymouse.h" />
+    <ClInclude Include="..\src\hardware\input\mouse_queue.h" />
     <ClInclude Include="..\src\hardware\mame\emu.h" />
     <ClInclude Include="..\src\hardware\mame\saa1099.h" />
     <ClInclude Include="..\src\hardware\mame\sn76496.h" />
-    <ClInclude Include="..\src\hardware\mouse\mouse_common.h" />
-    <ClInclude Include="..\src\hardware\mouse\mouse_config.h" />
-    <ClInclude Include="..\src\hardware\mouse\mouse_interfaces.h" />
-    <ClInclude Include="..\src\hardware\mouse\mouse_manymouse.h" />
-    <ClInclude Include="..\src\hardware\mouse\mouse_queue.h" />
     <ClInclude Include="..\src\hardware\serialport\directserial.h" />
     <ClInclude Include="..\src\hardware\serialport\libserial.h" />
     <ClInclude Include="..\src\hardware\serialport\misc_util.h" />
